{{cookiecutter.project_slug}}/requirements/local.txt
Werkzeug[watchdog]==3.1.8 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
{%- if cookiecutter.use_docker == 'y' %}
psycopg[c]==3.3.3  # https://github.com/psycopg/psycopg
{%- else %}
psycopg[binary]==3.3.3  # https://github.com/psycopg/psycopg
{%- endif %}
{%- if cookiecutter.use_async == 'y' or cookiecutter.use_celery == 'y' %}
watchfiles==1.1.1  # https://github.com/samuelcolvin/watchfiles
{%- endif %}

# Testing
# ------------------------------------------------------------------------------
mypy==1.20.0  # https://github.com/python/mypy
django-stubs[compatible-mypy]==6.0.2  # https://github.com/typeddjango/django-stubs
pytest==9.0.2  # https://github.com/pytest-dev/pytest
pytest-sugar==1.1.1  # https://github.com/Teemu/pytest-sugar
{%- if cookiecutter.rest_api == 'DRF' %}
djangorestframework-stubs==3.16.9  # https://github.com/typeddjango/djangorestframework-stubs
{%- endif %}

# Documentation
# ------------------------------------------------------------------------------
sphinx==9.1.0  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2025.8.25 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
ruff==0.15.9  # https://github.com/astral-sh/ruff
coverage==7.13.5  # https://github.com/nedbat/coveragepy
djlint==1.36.4  # https://github.com/Riverside-Healthcare/djLint
pre-commit==4.5.1  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.3  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==6.3.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==4.1  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.2.2  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.12.0  # https://github.com/pytest-dev/pytest-django


# Bdd
# ------------------------------------------------------------------------------
django-harlequin==1.5.0
harlequin[postgres]==2.0.5


# django-tailwind[reload]==3.8.0
# django-tailwind-4[reload]==0.1.4
